use-latest-values - move dog-gender into enums
