+++ src/lib/bark/enums/dog-gender.ts
