+++ src/lib/data/db-models.ts
@@ -16,7 +16,7 @@
 
 import { DogAntigenPresence } from "./db-enums";
 import { YesNoUnknown } from "../bark/enums/yes-no-unknown";
-import { DogGender } from "../bark/models/dog-gender";
+import { DogGender } from "../bark/enums/dog-gender";
 import { UserResidency } from "./db-enums";
 import { CallOutcome, PosNegNil, ReportedIneligibility } from "./db-enums";
