+++ tests/admin/get-incomplete-profiles.test.ts
@@ -11,7 +11,7 @@ import { getIncompleteProfiles } from "@/lib/admin/actions/get-incomplete-profil
 import { Pool } from "pg";
 import { PARTICIPATION_STATUS } from "@/lib/data/db-enums";
 import { YES_NO_UNKNOWN } from "@/lib/bark/enums/yes-no-unknown";
-import { DOG_GENDER } from "@/lib/bark/models/dog-gender";
+import { DOG_GENDER } from "@/lib/bark/enums/dog-gender";
 import {
   DEFAULT_DATE_TIME_FORMAT,
   SINGAPORE_TIME_ZONE,
