+++ tests/user/get-dog-statuses.test.ts
@@ -9,7 +9,7 @@ import {
   SERVICE_STATUS,
 } from "@/lib/data/db-enums";
 import { YES_NO_UNKNOWN } from "@/lib/bark/enums/yes-no-unknown";
-import { DOG_GENDER } from "@/lib/bark/models/dog-gender";
+import { DOG_GENDER } from "@/lib/bark/enums/dog-gender";
 import { dateAgo } from "../_time_helpers";
 
 describe("getDogStatuses", () => {
