+++ tests/user/update-dog-profile.test.ts
@@ -13,7 +13,7 @@ import { DogProfile } from "@/lib/dog/dog-models";
 import { UTC_DATE_OPTION, parseDateTime } from "@/lib/utilities/bark-time";
 import { CALL_OUTCOME, DOG_ANTIGEN_PRESENCE } from "@/lib/data/db-enums";
 import { YES_NO_UNKNOWN } from "@/lib/bark/enums/yes-no-unknown";
-import { DOG_GENDER } from "@/lib/bark/models/dog-gender";
+import { DOG_GENDER } from "@/lib/bark/enums/dog-gender";
 import { dbInsertDogVetPreference } from "@/lib/data/db-dogs";
 
 describe("updateDogProfile", () => {
